@@ -1,3 +1,3 @@
 module github.com/piotrnar/gocoin
 
-go 1.16
+go 1.18
