@@ -3,8 +3,8 @@ package main
 import (
 	"crypto/rand"
 	"fmt"
-	"io/ioutil"
 	"os"
+	"sync"
 	"time"
 
 	"github.com/piotrnar/gocoin/client/common"
@@ -38,17 +38,18 @@ func host_init() {
 	os.MkdirAll(common.GocoinHomeDir, 0770)
 	sys.LockDatabaseDir(common.GocoinHomeDir)
 
-	common.SecretKey, _ = ioutil.ReadFile(common.GocoinHomeDir + "authkey")
+	common.SecretKey, _ = os.ReadFile(common.GocoinHomeDir + "authkey")
 	if len(common.SecretKey) != 32 {
 		common.SecretKey = make([]byte, 32)
 		rand.Read(common.SecretKey)
-		ioutil.WriteFile(common.GocoinHomeDir+"authkey", common.SecretKey, 0600)
+		os.WriteFile(common.GocoinHomeDir+"authkey", common.SecretKey, 0600)
 	}
 	common.PublicKey = btc.Encodeb58(btc.PublicFromPrivate(common.SecretKey, true))
 	fmt.Println("Public auth key:", "@"+common.PublicKey)
 
 	__exit := make(chan bool)
-	__done := make(chan bool)
+	var done sync.WaitGroup
+	done.Add(1)
 	go func() {
 		for {
 			select {
@@ -56,7 +57,7 @@ func host_init() {
 				fmt.Println(s)
 				chain.AbortNow = true
 			case <-__exit:
-				__done <- true
+				done.Done()
 				return
 			}
 		}
@@ -67,7 +68,7 @@ func host_init() {
 		os.Exit(1)
 	}
 
-	fmt.Print(string(common.LogBuffer.Bytes()))
+	fmt.Print(common.LogBuffer.String())
 	common.LogBuffer = nil
 
 	if btc.EC_Verify == nil {
@@ -75,6 +76,7 @@ func host_init() {
 	}
 
 	ext := &chain.NewChanOpts{
+		UtxoFilesSubdir:  common.CFG.UtxoSubdir,
 		UTXOVolatileMode: common.FLAG.VolatileUTXO,
 		UndoBlocks:       common.FLAG.UndoBlocks,
 		BlockMinedCB:     blockMined, BlockUndoneCB: blockUndone,
@@ -99,11 +101,13 @@ func host_init() {
 			DataFilesBackup: common.CFG.Memory.OldDataBackup,
 			CompressOnDisk:  common.CFG.Memory.CompressBlockDB})
 	if chain.AbortNow {
-		fmt.Printf("Blockchain opening aborted after %s seconds\n", time.Now().Sub(sta).String())
+		fmt.Printf("Blockchain opening aborted after %s seconds\n", time.Since(sta).String())
 		common.BlockChain.Close()
 		sys.UnlockDatabaseDir()
 		os.Exit(1)
 	}
+	__exit <- true
+	done.Wait()
 
 	if lb, _ := common.BlockChain.BlockTreeRoot.FindFarthestNode(); lb.Height > common.BlockChain.LastBlock().Height {
 		common.Last.ParseTill = lb
@@ -135,7 +139,5 @@ func host_init() {
 		common.Last.Time.Format("2006-01-02 15:04:05"))
 
 	common.StartTime = time.Now()
-	__exit <- true
-	_ = <-__done
 
 }
