@@ -72,7 +72,7 @@ func (c *OneConnection) ProcessNewHeader(hdr []byte) (int, *OneBlockToGet) {
 	}
 
 	if common.LastTrustedBlockMatch(node.BlockHash) {
-		common.SetUint32(&common.LastTrustedBlockHeight, node.Height)
+		common.Set(&common.LastTrustedBlockHeight, node.Height)
 		for node != nil {
 			node.Trusted.Set()
 			node = node.Parent
@@ -255,8 +255,6 @@ func (c *OneConnection) GetHeaders(pl []byte) {
 	}
 
 	// Note: the deferred function will be called before exiting
-
-	return
 }
 
 func (c *OneConnection) sendGetHeaders() {
