@@ -1,8 +1,6 @@
-package network
+package txpool
 
 import (
-	"bytes"
-	"errors"
 	"fmt"
 
 	"github.com/piotrnar/gocoin/client/common"
@@ -26,7 +24,7 @@ func (tx *OneTxToSend) UnMarkChildrenForMem() {
 	for po.Vout = 0; po.Vout < uint32(len(tx.TxOut)); po.Vout++ {
 		uidx := po.UIdx()
 		if val, ok := SpentOutputs[uidx]; ok {
-			if rec, _ := TransactionsToSend[val]; rec != nil {
+			if rec := TransactionsToSend[val]; rec != nil {
 				if rec.MemInputs == nil {
 					common.CountSafe("TxMinedMeminER1")
 					fmt.Println("WTF?", po.String(), "just mined in", rec.Hash.String(), "- not marked as mem")
@@ -43,8 +41,12 @@ func (tx *OneTxToSend) UnMarkChildrenForMem() {
 				common.CountSafe("TxMinedMeminOut")
 				if rec.MemInputCnt == 0 {
 					common.CountSafe("TxMinedMeminTx")
+					reduxed_size := (len(rec.MemInputs) + 7) & ^7
 					rec.MemInputs = nil
+					rec.Footprint -= uint32(reduxed_size)
+					TransactionsToSendSize -= uint64(reduxed_size)
 				}
+				rec.ResortWithChildren()
 			} else {
 				common.CountSafe("TxMinedMeminERR")
 				fmt.Println("WTF?", po.String(), " in SpentOutputs, but not in mempool")
@@ -54,31 +56,29 @@ func (tx *OneTxToSend) UnMarkChildrenForMem() {
 }
 
 // tx_mined is called for each tx mined in a new block.
-func tx_mined(tx *btc.Tx) (wtg *OneWaitingList) {
+func tx_mined(tx *btc.Tx) {
 	h := tx.Hash
 	if rec, ok := TransactionsToSend[h.BIdx()]; ok {
-		common.CountSafe("TxMinedToSend")
+		common.CountSafe("TxMinedAccepted")
 		rec.UnMarkChildrenForMem()
 		rec.Delete(false, 0)
 	}
 	if mr, ok := TransactionsRejected[h.BIdx()]; ok {
-		if mr.Tx != nil {
-			common.CountSafe(fmt.Sprint("TxMinedROK-", mr.Reason))
-		} else {
-			common.CountSafe(fmt.Sprint("TxMinedRNO-", mr.Reason))
-		}
-		deleteRejected(h.BIdx())
+		common.CountSafePar("TxMinedRejected-", mr.Reason)
+		DeleteRejectedByTxr(mr)
 	}
 	if _, ok := TransactionsPending[h.BIdx()]; ok {
 		common.CountSafe("TxMinedPending")
 		delete(TransactionsPending, h.BIdx())
 	}
 
-	// Go through all the inputs and make sure we are not leaving them in SpentOutputs
-	for i := range tx.TxIn {
-		idx := tx.TxIn[i].Input.UIdx()
+	// now do through all the spent inputs and...
+	for _, inp := range tx.TxIn {
+		idx := inp.Input.UIdx()
+
+		// 1. make sure we are not leaving them in SpentOutputs
 		if val, ok := SpentOutputs[idx]; ok {
-			if rec, _ := TransactionsToSend[val]; rec != nil {
+			if rec := TransactionsToSend[val]; rec != nil {
 				// if we got here, the txs has been Malleabled
 				if rec.Local {
 					common.CountSafe("TxMinedMalleabled")
@@ -88,40 +88,48 @@ func tx_mined(tx *btc.Tx) (wtg *OneWaitingList) {
 				}
 				rec.Delete(true, 0)
 			} else {
-				common.CountSafe("TxMinedSpentERROR")
-				fmt.Println("WTF? Input from ", rec.Tx.Hash.String(), " in mem-spent, but tx not in the mem-pool")
+				println("ERROR: Input from ", inp.Input.String(), " in SpentOutputs, but tx not in mempool")
 			}
 			delete(SpentOutputs, idx)
 		}
-	}
 
-	wtg = WaitingForInputs[h.BIdx()]
-	return
+		// 2. remove data of any rejected txs that use this input
+		if lst, ok := RejectedUsedUTXOs[idx]; ok {
+			for _, bidx := range lst {
+				if txr, ok := TransactionsRejected[bidx]; ok {
+					common.CountSafePar("TxMinedRjctUTXO-", txr.Reason)
+					DeleteRejectedByTxr(txr)
+				} else {
+					println("ERROR: txr marked for removal but not present in TransactionsRejected")
+				}
+			}
+			delete(RejectedUsedUTXOs, idx) // this record will not be needed anymore
+		}
+	}
 }
 
 // BlockMined removes all the block's tx from the mempool.
 func BlockMined(bl *btc.Block) {
-	wtgs := make([]*OneWaitingList, len(bl.Txs)-1)
-	var wtg_cnt int
+	wtgs := make([]*OneWaitingList, 0, len(bl.Txs)-1)
 	TxMutex.Lock()
-	for i := 1; i < len(bl.Txs); i++ {
-		wtg := tx_mined(bl.Txs[i])
-		if wtg != nil {
-			wtgs[wtg_cnt] = wtg
-			wtg_cnt++
+	for _, tx := range bl.Txs[1:] {
+		tx_mined(tx)
+	}
+	for _, tx := range bl.Txs[1:] {
+		bidx := tx.Hash.BIdx()
+		if wtg := WaitingForInputs[bidx]; wtg != nil {
+			wtgs = append(wtgs, wtg)
 		}
 	}
 	TxMutex.Unlock()
 
 	// Try to redo waiting txs
-	if wtg_cnt > 0 {
-		common.CountSafeAdd("TxMinedGotInput", uint64(wtg_cnt))
-		for _, wtg := range wtgs[:wtg_cnt] {
+	if len(wtgs) > 0 {
+		common.CountSafeAdd("TxMinedGotInput", uint64(len(wtgs)))
+		for _, wtg := range wtgs {
 			RetryWaitingForInput(wtg)
 		}
 	}
-
-	expireTxsNow = true
 }
 
 // MarkChildrenForMem sets the MemInput flag of all the children (used when a tx is mined).
@@ -132,17 +140,18 @@ func MarkChildrenForMem(tx *btc.Tx) {
 	for po.Vout = 0; po.Vout < uint32(len(tx.TxOut)); po.Vout++ {
 		uidx := po.UIdx()
 		if val, ok := SpentOutputs[uidx]; ok {
-			if rec, _ := TransactionsToSend[val]; rec != nil {
+			if rec := TransactionsToSend[val]; rec != nil {
 				if rec.MemInputs == nil {
 					rec.MemInputs = make([]bool, len(rec.TxIn))
 				}
 				idx := rec.IIdx(uidx)
 				rec.MemInputs[idx] = true
 				rec.MemInputCnt++
+				rec.ResortWithChildren()
 				common.CountSafe("TxPutBackMemIn")
 			} else {
+				println("ERROR: MarkChildrenForMem", po.String(), " in SpentOutputs, but not in mempool")
 				common.CountSafe("TxPutBackMeminERR")
-				fmt.Println("MarkChildrenForMem WTF?", po.String(), " in SpentOutputs, but not in mempool")
 			}
 		}
 	}
@@ -152,7 +161,7 @@ func BlockUndone(bl *btc.Block) {
 	var cnt int
 	for _, tx := range bl.Txs[1:] {
 		// put it back into the mempool
-		ntx := &TxRcvd{Tx: tx, trusted: true}
+		ntx := &TxRcvd{Tx: tx, Trusted: true}
 
 		if NeedThisTx(&ntx.Hash, nil) {
 			if HandleNetTx(ntx, true) {
@@ -165,93 +174,9 @@ func BlockUndone(bl *btc.Block) {
 			common.CountSafe("TxPutBackNoNeed")
 		}
 
-		// TODO: make sure to set MemInputs of ones using it back to true (issue #58)
 		MarkChildrenForMem(tx)
 	}
 	if cnt != len(bl.Txs)-1 {
 		println("WARNING: network.BlockUndone("+bl.Hash.String()+") - ", cnt, "of", len(bl.Txs)-1, "txs put back")
 	}
-}
-
-func (c *OneConnection) SendGetMP() error {
-	if len(c.GetMP) == 0 {
-		// TODO: remove it at some point (should not be happening)
-		println("ERROR: SendGetMP() called with no GetMP lock")
-		return nil
-	}
-	b := new(bytes.Buffer)
-	TxMutex.Lock()
-	if TransactionsToSendSize > common.MaxMempoolSize()>>1 {
-		// Don't send "getmp" messages if we have more than 50% of MaxMempoolSize() used
-		//fmt.Println("Mempool more than half full - not sending getmp message -", TransactionsToSendSize>>20, "/", common.MaxMempoolSize()>>20)
-		TxMutex.Unlock()
-		c.cntInc("GetMPHold")
-		return errors.New("SendGetMP: Mempool more than half full")
-	}
-	tcnt := len(TransactionsToSend) + len(TransactionsRejected)
-	if tcnt > MAX_GETMP_TXS {
-		fmt.Println("Too many transactions in the current pool", tcnt, "/", MAX_GETMP_TXS)
-		tcnt = MAX_GETMP_TXS
-	}
-	btc.WriteVlen(b, uint64(tcnt))
-	var cnt int
-	for k := range TransactionsToSend {
-		b.Write(k[:])
-		cnt++
-		if cnt == MAX_GETMP_TXS {
-			break
-		}
-	}
-	for k := range TransactionsRejected {
-		b.Write(k[:])
-		cnt++
-		if cnt == MAX_GETMP_TXS {
-			break
-		}
-	}
-	TxMutex.Unlock()
-	return c.SendRawMsg("getmp", b.Bytes())
-}
-
-func (c *OneConnection) ProcessGetMP(pl []byte) {
-	br := bytes.NewBuffer(pl)
-
-	cnt, er := btc.ReadVLen(br)
-	if er != nil {
-		println("getmp message does not have the length field")
-		c.DoS("GetMPError1")
-		return
-	}
-
-	has_this_one := make(map[BIDX]bool, cnt)
-	for i := 0; i < int(cnt); i++ {
-		var idx BIDX
-		if n, _ := br.Read(idx[:]); n != len(idx) {
-			println("getmp message too short")
-			c.DoS("GetMPError2")
-			return
-		}
-		has_this_one[idx] = true
-	}
-
-	var data_sent_so_far int
-	var redo [1]byte
-
-	TxMutex.Lock()
-	for k, v := range TransactionsToSend {
-		c.Mutex.Lock()
-		bts := c.BytesToSent()
-		c.Mutex.Unlock()
-		if bts > SendBufSize/4 {
-			redo[0] = 1
-			break
-		}
-		if !has_this_one[k] {
-			c.SendRawMsg("tx", v.Raw)
-			data_sent_so_far += 24 + len(v.Raw)
-		}
-	}
-	TxMutex.Unlock()
-
-	c.SendRawMsg("getmpdone", redo[:])
 }
