@@ -15,30 +15,28 @@ var (
 )
 
 func InitMaps(empty bool) {
-	var szs [5]int
+	var szs [IDX_CNT]int
 
 	if !empty {
 		LoadMapSizes()
-		szs, _ = WalletAddrsCount[common.AllBalMinVal()]
+		szs = WalletAddrsCount[common.AllBalMinVal()]
 		// If yet unknown, just continue with zero size maps
 	}
 
-	AllBalancesP2KH = make(map[[20]byte]*OneAllAddrBal, szs[0])
-	AllBalancesP2SH = make(map[[20]byte]*OneAllAddrBal, szs[1])
-	AllBalancesP2WKH = make(map[[20]byte]*OneAllAddrBal, szs[2])
-	AllBalancesP2WSH = make(map[[32]byte]*OneAllAddrBal, szs[3])
-	AllBalancesP2TAP = make(map[[32]byte]*OneAllAddrBal, szs[4])
+	for i := range AllBalances {
+		AllBalances[i] = make(map[string]*OneAllAddrBal, szs[i])
+	}
 }
 
-func LoadBalance() {
-	if common.GetBool(&common.WalletON) {
+func LoadBalancesFromUtxo() {
+	if common.Get(&common.WalletON) {
 		//fmt.Println("wallet.LoadBalance() ignore: ", common.GetBool(&common.WalletON))
 		return
 	}
 
 	var aborted bool
 
-	common.SetUint32(&common.WalletProgress, 1)
+	common.Set(&common.WalletProgress, 1)
 	common.ApplyBalMinVal()
 
 	InitMaps(false)
@@ -46,7 +44,7 @@ func LoadBalance() {
 	for _i := range common.BlockChain.Unspent.HashMap {
 		common.BlockChain.Unspent.MapMutex[_i].RLock()
 		for k, v := range common.BlockChain.Unspent.HashMap[_i] {
-			NewUTXO(utxo.NewUtxoRecStatic(k, v))
+			TxNotifyAdd(utxo.NewUtxoRecStatic(k, v))
 			if FetchingBalanceTick != nil && FetchingBalanceTick() {
 				aborted = true
 				break
@@ -56,27 +54,27 @@ func LoadBalance() {
 		if aborted {
 			break
 		}
-		common.SetUint32(&common.WalletProgress, 1000*(uint32(_i)+1)/256)
+		common.Set(&common.WalletProgress, 1000*(uint32(_i)+1)/256)
 	}
 	if aborted {
 		InitMaps(true)
 	} else {
 		common.BlockChain.Unspent.CB.NotifyTxAdd = TxNotifyAdd
 		common.BlockChain.Unspent.CB.NotifyTxDel = TxNotifyDel
-		common.SetBool(&common.WalletON, true)
+		common.Set(&common.WalletON, true)
 	}
-	common.SetUint32(&common.WalletProgress, 0)
+	common.Set(&common.WalletProgress, 0)
 }
 
 func Disable() {
-	if !common.GetBool(&common.WalletON) {
+	if !common.Get(&common.WalletON) {
 		//fmt.Println("wallet.Disable() ignore: ", common.GetBool(&common.WalletON))
 		return
 	}
 	UpdateMapSizes()
 	common.BlockChain.Unspent.CB.NotifyTxAdd = nil
 	common.BlockChain.Unspent.CB.NotifyTxDel = nil
-	common.SetBool(&common.WalletON, false)
+	common.Set(&common.WalletON, false)
 	InitMaps(true)
 }
 
@@ -85,14 +83,15 @@ const (
 )
 
 var (
-	WalletAddrsCount map[uint64][5]int = make(map[uint64][5]int) //index:MinValue, [0]-P2KH, [1]-P2SH, [2]-P2WSH, [3]-P2WKH, [4]-P2TAP
+	WalletAddrsCount map[uint64][IDX_CNT]int = make(map[uint64][IDX_CNT]int) //index:MinValue, [0]-P2KH, [1]-P2SH, [2]-P2WSH, [3]-P2WKH, [4]-P2TAP
 )
 
 func UpdateMapSizes() {
-	WalletAddrsCount[common.AllBalMinVal()] = [5]int{len(AllBalancesP2KH),
-		len(AllBalancesP2SH), len(AllBalancesP2WKH), len(AllBalancesP2WSH),
-		len(AllBalancesP2TAP)}
-
+	var tmp [IDX_CNT]int
+	for i := range tmp {
+		tmp[i] = len(AllBalances[i])
+	}
+	WalletAddrsCount[common.AllBalMinVal()] = tmp
 	buf := new(bytes.Buffer)
 	gob.NewEncoder(buf).Encode(WalletAddrsCount)
 	os.WriteFile(common.GocoinHomeDir+MAPSIZ_FILE_NAME, buf.Bytes(), 0600)
