@@ -1,79 +1,12 @@
 package utils
 
 import (
-	"encoding/hex"
-	"encoding/json"
 	"fmt"
-	"github.com/piotrnar/gocoin/lib/btc"
 	"io"
 	"net/http"
+
+	"github.com/piotrnar/gocoin/lib/btc"
 )
-
-// https://blockchain.info/block/000000000000000000871f4f01a389bda59e568ead8d0fd45fc7cc1919d2666e?format=hex
-// https://webbtc.com/block/0000000000000000000cdc0d2a9b33c2d4b34b4d4fa8920f074338d0dc1164dc.bin
-// https://blockexplorer.com/api/rawblock/0000000000000000000cdc0d2a9b33c2d4b34b4d4fa8920f074338d0dc1164dc
-
-// GetBlockFromExplorer downloads (and re-assembles) a raw block from blockexplorer.com.
-func GetBlockFromExplorer(hash *btc.Uint256) (rawtx []byte) {
-	url := "http://blockexplorer.com/api/rawblock/" + hash.String()
-	r, er := http.Get(url)
-	if er == nil {
-		if r.StatusCode == 200 {
-			defer r.Body.Close()
-			c, _ := io.ReadAll(r.Body)
-			var txx struct {
-				Raw string `json:"rawblock"`
-			}
-			er = json.Unmarshal(c[:], &txx)
-			if er == nil {
-				rawtx, er = hex.DecodeString(txx.Raw)
-			}
-		} else {
-			fmt.Println("blockexplorer.com StatusCode=", r.StatusCode)
-		}
-	}
-	if er != nil {
-		fmt.Println("blockexplorer.com:", er.Error())
-	}
-	return
-}
-
-// GetBlockFromWebBTC downloads a raw block from webbtc.com.
-func GetBlockFromWebBTC(hash *btc.Uint256) (raw []byte) {
-	url := "https://webbtc.com/block/" + hash.String() + ".bin"
-	r, er := http.Get(url)
-	if er == nil {
-		if r.StatusCode == 200 {
-			raw, _ = io.ReadAll(r.Body)
-			r.Body.Close()
-		} else {
-			fmt.Println("webbtc.com StatusCode=", r.StatusCode)
-		}
-	}
-	if er != nil {
-		fmt.Println("webbtc.com:", er.Error())
-	}
-	return
-}
-
-// GetBlockFromBlockchainInfo downloads a raw block from blockchain.info.
-func GetBlockFromBlockchainInfo(hash *btc.Uint256) (rawtx []byte) {
-	url := "https://blockchain.info/block/" + hash.String() + "?format=hex"
-	r, er := http.Get(url)
-	if er == nil {
-		if r.StatusCode == 200 {
-			defer r.Body.Close()
-			rawhex, _ := io.ReadAll(r.Body)
-			rawtx, er = hex.DecodeString(string(rawhex))
-		} else {
-			fmt.Println("blockchain.info StatusCode=", r.StatusCode)
-		}
-	}
-	if er != nil {
-		fmt.Println("blockchain.info:", er.Error())
-	}
-	return
-}
 
 // GetBlockFromBlockstream downloads a raw block from blockstream
 func GetBlockFromBlockstream(hash *btc.Uint256, api_url string) (raw []byte) {
@@ -118,26 +51,24 @@ func GetBlockFromWeb(hash *btc.Uint256) (bl *btc.Block) {
 
 	raw = GetBlockFromBlockstream(hash, "https://blockstream.info/api/block/")
 	if bl = IsBlockOK(raw, hash); bl != nil {
-		//println("GetTxFromBlockstream - OK")
+		if Verbose {
+			println("GetTxFromBlockstream - OK")
+		}
 		return
 	}
-
-	raw = GetBlockFromBlockchainInfo(hash)
-	if bl = IsBlockOK(raw, hash); bl != nil {
-		//println("GetTxFromBlockchainInfo - OK")
-		return
+	if Verbose {
+		println("GetTxFromBlockstream error")
 	}
 
-	raw = GetBlockFromExplorer(hash)
+	raw = GetBlockFromBlockstream(hash, "https://mempool.space/api/block/")
 	if bl = IsBlockOK(raw, hash); bl != nil {
-		//println("GetTxFromExplorer - OK")
+		if Verbose {
+			println("GetBlockFromMempoolSpace - OK")
+		}
 		return
 	}
-
-	raw = GetBlockFromWebBTC(hash)
-	if bl = IsBlockOK(raw, hash); bl != nil {
-		//println("GetTxFromWebBTC - OK")
-		return
+	if Verbose {
+		println("GetBlockFromMempoolSpace error")
 	}
 
 	return
