@@ -6,174 +6,11 @@ import (
 	"fmt"
 	"io"
 	"net/http"
-	"os"
 	"time"
 
 	"github.com/piotrnar/gocoin/lib/btc"
 	"github.com/piotrnar/gocoin/lib/utxo"
 )
-
-func GetUnspentFromExplorer(addr *btc.BtcAddr, testnet bool) (res utxo.AllUnspentTx, er error) {
-	var r *http.Response
-	if testnet {
-		r, er = http.Get("https://testnet.blockexplorer.com/api/addr/" + addr.String() + "/utxo")
-	} else {
-		r, er = http.Get("https://blockexplorer.com/api/addr/" + addr.String() + "/utxo")
-	}
-	if er != nil {
-		return
-	}
-	if r.StatusCode != 200 {
-		er = errors.New(fmt.Sprint("HTTP StatusCode ", r.StatusCode))
-		return
-	}
-
-	c, _ := io.ReadAll(r.Body)
-	r.Body.Close()
-
-	var result []struct {
-		Addr   string `json:"address"`
-		TxID   string `json:"txid"`
-		Vout   uint32 `json:"vout"`
-		Value  uint64 `json:"satoshis"`
-		Height uint32 `json:"height"`
-	}
-
-	er = json.Unmarshal(c, &result)
-	if er != nil {
-		return
-	}
-
-	for _, r := range result {
-		ur := new(utxo.OneUnspentTx)
-		id := btc.NewUint256FromString(r.TxID)
-		if id == nil {
-			er = errors.New(fmt.Sprint("Bad TXID:", r.TxID))
-			return
-		}
-		copy(ur.TxPrevOut.Hash[:], id.Hash[:])
-		ur.TxPrevOut.Vout = r.Vout
-		ur.Value = r.Value
-		ur.MinedAt = r.Height
-		ur.BtcAddr = addr
-		res = append(res, ur)
-	}
-
-	return
-}
-
-func GetUnspentFromBlockchainInfo(addr *btc.BtcAddr) (res utxo.AllUnspentTx, er error) {
-	var r *http.Response
-	r, er = http.Get("https://blockchain.info/unspent?active=" + addr.String())
-	if er != nil {
-		return
-	}
-	if r.StatusCode != 200 {
-		er = errors.New(fmt.Sprint("HTTP StatusCode ", r.StatusCode))
-		return
-	}
-
-	c, _ := io.ReadAll(r.Body)
-	r.Body.Close()
-
-	var result struct {
-		U []struct {
-			TxID  string `json:"tx_hash_big_endian"`
-			Vout  uint32 `json:"tx_output_n"`
-			Value uint64 `json:"value"`
-		} `json:"unspent_outputs"`
-	}
-
-	er = json.Unmarshal(c, &result)
-	if er != nil {
-		return
-	}
-
-	for _, r := range result.U {
-		ur := new(utxo.OneUnspentTx)
-		id := btc.NewUint256FromString(r.TxID)
-		if id == nil {
-			er = errors.New(fmt.Sprint("Bad TXID:", r.TxID))
-			return
-		}
-		copy(ur.TxPrevOut.Hash[:], id.Hash[:])
-		ur.TxPrevOut.Vout = r.Vout
-		ur.Value = r.Value
-		//ur.MinedAt = r.Height
-		ur.BtcAddr = addr
-		res = append(res, ur)
-	}
-
-	return
-}
-
-func GetUnspentFromBlockcypher(addr *btc.BtcAddr, currency string) (res utxo.AllUnspentTx, er error) {
-	var r *http.Response
-	var try_cnt int
-
-	token := os.Getenv("BLOCKCYPHER_TOKEN")
-	if token == "" {
-		println("WARNING: BLOCKCYPHER_TOKEN envirionment variable not set (get it from blockcypher.com)")
-	} else {
-		token = "&token=" + token
-	}
-
-	for {
-		r, er = http.Get("https://api.blockcypher.com/v1/" + currency + "/main/addrs/" + addr.String() + "?unspentOnly=true" + token)
-
-		if er != nil {
-			return
-		}
-		if r.StatusCode == 429 && try_cnt < 5 {
-			try_cnt++
-			println("Retry blockcypher.com in", try_cnt, "seconds...")
-			time.Sleep(time.Duration(try_cnt) * time.Second)
-			continue
-		}
-
-		if r.StatusCode != 200 {
-			er = errors.New(fmt.Sprint("HTTP StatusCode ", r.StatusCode))
-			return
-		}
-
-		break
-	}
-
-	c, _ := io.ReadAll(r.Body)
-	r.Body.Close()
-
-	var result struct {
-		Addr string `json:"address"`
-		Outs []struct {
-			TxID   string `json:"tx_hash"`
-			Vout   uint32 `json:"tx_output_n"`
-			Value  uint64 `json:"value"`
-			Height uint32 `json:"block_height"`
-		} `json:"txrefs"`
-	}
-
-	er = json.Unmarshal(c, &result)
-	if er != nil {
-		return
-	}
-
-	for _, r := range result.Outs {
-		ur := new(utxo.OneUnspentTx)
-		id := btc.NewUint256FromString(r.TxID)
-		if id == nil {
-			er = errors.New(fmt.Sprint("Bad TXID:", r.TxID))
-			return
-		}
-		copy(ur.TxPrevOut.Hash[:], id.Hash[:])
-		ur.TxPrevOut.Vout = r.Vout
-		ur.Value = r.Value
-		ur.MinedAt = r.Height
-		ur.BtcAddr = addr
-		res = append(res, ur)
-	}
-
-	return
-}
 
 // currency is either "bitcoin" or "bitcoin-cash"
 func GetUnspentFromBlockchair(addr *btc.BtcAddr, currency string) (res utxo.AllUnspentTx, er error) {
@@ -296,33 +133,37 @@ func GetUnspent(addr *btc.BtcAddr) (res utxo.AllUnspentTx) {
 
 	res, er = GetUnspentFromBlockstream(addr, "https://blockstream.info/api/address/")
 	if er == nil {
+		if Verbose {
+			println("GetUnspentFromBlockstream OK")
+		}
 		return
 	}
-	println("GetUnspentFromBlockstream:", er.Error())
+	if Verbose {
+		println("GetUnspentFromBlockstream:", er.Error())
+	}
+
+	res, er = GetUnspentFromBlockstream(addr, "https://mempool.space/api/address/")
+	if er == nil {
+		if Verbose {
+			println("GetUnspentFromMempoolSpace OK")
+		}
+		return
+	}
+	if Verbose {
+		println("GetUnspentFromMempoolSpace:", er.Error())
+	}
 
 	res, er = GetUnspentFromBlockchair(addr, "bitcoin")
 	if er == nil {
+		if Verbose {
+			println("GetUnspentFromBlockchair OK")
+		}
 		return
 	}
-	println("GetUnspentFromBlockchair:", er.Error())
+	if Verbose {
+		println("GetUnspentFromBlockchair:", er.Error())
+	}
 
-	res, er = GetUnspentFromBlockcypher(addr, "btc")
-	if er == nil {
-		return
-	}
-	println("GetUnspentFromBlockcypher:", er.Error())
-
-	res, er = GetUnspentFromBlockchainInfo(addr)
-	if er == nil {
-		return
-	}
-	println("GetUnspentFromBlockchainInfo:", er.Error())
-
-	res, er = GetUnspentFromExplorer(addr, false)
-	if er == nil {
-		return
-	}
-	println("GetUnspentFromExplorer:", er.Error())
 	return
 }
 
@@ -331,21 +172,25 @@ func GetUnspentTestnet(addr *btc.BtcAddr) (res utxo.AllUnspentTx) {
 
 	res, er = GetUnspentFromBlockstream(addr, "https://blockstream.info/testnet/api/address/")
 	if er == nil {
+		if Verbose {
+			println("GetUnspentFromBlockstream OK")
+		}
 		return
 	}
-	println("GetUnspentFromBlockstream:", er.Error())
+	if Verbose {
+		println("GetUnspentFromBlockstream:", er.Error())
+	}
 
-	res, er = GetUnspentFromExplorer(addr, true)
+	res, er = GetUnspentFromBlockstream(addr, "https://mempool.space/testnet/api/address/")
 	if er == nil {
+		if Verbose {
+			println("GetUnspentFromMempoolSpace OK")
+		}
 		return
 	}
-	println("GetUnspentFromExplorer:", er.Error())
-
-	res, er = GetUnspentFromBlockcypher(addr, "btc-testnet")
-	if er == nil {
-		return
+	if Verbose {
+		println("GetUnspentFromMempoolSpace:", er.Error())
 	}
-	println("GetUnspentFromBlockcypher:", er.Error())
 
 	return
 }
